requirements.txt
# FastAPI and backend dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3

# Scientific computing packages (user specified versions)
netCDF4==1.7.2
numpy==2.1.3
pandas==2.2.3
xarray==2025.3.0
statsmodels==0.14.4
dask==2025.3.0
cdsapi==0.7.5
matplotlib
xesmf==0.8.1
gdown==5.2.0
scipy==1.15.2

# R integration
rpy2==3.6.0

# Database and caching (optional)
redis==5.2.1
celery==5.4.0
sqlalchemy[asyncio]
asyncpg
alembic==1.14.0
psycopg2-binary==2.9.10

# Security (for future authentication)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
